v1.2.11
- Update dependencies
- Split some parts to utils
