+++ package.json
@@ -1,5 +1,5 @@
 {
-  "version": "1.2.1",
+  "version": "1.2.11",
   "name": "bella-scheduler",
   "description": "To do something later, repeater",
   "homepage": "https://www.npmjs.com/package/bella-scheduler",
@@ -18,7 +18,7 @@
   "scripts": {
     "lint": "eslint src tests",
     "nsp": "nsp check --output checkstyle",
-    "tape": "nyc babel-node node_modules/.bin/tape tests/start.js | tap-spec",
+    "tape": "nyc babel-node node_modules/.bin/tape test/start.js | tap-spec",
     "report": "nyc report --reporter=lcov > coverage.lcov && codecov",
     "pretest": "npm run lint && npm run nsp",
     "test": "npm run tape",
@@ -35,7 +35,7 @@
     "babel-plugin-transform-remove-strict-mode": "0.x.x",
     "babel-preset-env": "1.x.x",
     "codecov": "2.x.x",
-    "eslint": "3.x.x",
+    "eslint": "4.x.x",
     "eslint-config-ggc": "1.x.x",
     "gcc-min": "5.x.x",
     "nsp": "2.x.x",
