+++ src/utils/index.js
@@ -1,0 +1,4 @@
+export * from './map';
+export * from './getDT1';
+export * from './getDT2';
+export * from './getDT3';
