+++ test/start.js
@@ -5,7 +5,7 @@ var path = require('path');
  * Import specs
  */
 
-var where = './tests/specs';
+var where = './test/specs';
 if (fs.existsSync(where)) {
   fs.readdirSync(where).forEach((file) => {
     if (path.extname(file) === '.js') {
